+++ src/processors/types/collection.ts
